alterando a tipagem
    Foi alterado a tipagem nos métodos setWidth e setHeight de float
    para int porque ele depende da biblioteca \Picqer\Barcode\BarcodeGeneratorPNG
    e nela o parâmetro dever int.

+++ src/Tags/Barcode.php
@@ -39,7 +39,7 @@ class Barcode
      * @param float $width
      * @return $this
      */
-    public function setWidth(float $width)
+    public function setWidth(int $width)
     {
         $this->width = $width;
         return $this;
@@ -49,7 +49,7 @@ class Barcode
      * @param float $height
      * @return $this
      */
-    public function setHeight(float $height)
+    public function setHeight(int $height)
     {
         $this->height = $height;
         return $this;
